More precize coordinate parsing now allowed for interchange paths

+++ choo/apis/efa.py
@@ -635,7 +635,7 @@ class EFA(API):
         path = []
         for coords in data.findall('./itdInterchangePathCoordinates/itdPathCoordinates/'
                                    '/itdCoordinateBaseElemList/itdCoordinateBaseElem'):
-            path.append(Coordinates(int(coords.find('y').text) / 1000000, int(coords.find('x').text) / 1000000))
+            path.append(Coordinates(float(coords.find('y').text) / 1000000, float(coords.find('x').text) / 1000000))
 
         if path:
             way.path = path
